Run tests on a future Django support branch
This is intended to make gradually adding support for future Django
versions easier. It creates a separate Travis job which merges
a future Django support branch into the current branch and runs
tests, and also runs a custom "linter" script which checks for
adding uses of features that won't work anymore and might not
have test coverage but are greppable.